requirements.txt
asgiref==3.5.2
autopep8==1.6.0
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
filelock==3.7.1
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
virtualenv==20.15.0
